requirements.txt
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
markdown-it-py==4.0.0
mdurl==0.1.2
pathlib==1.0.1
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
rich==14.2.0
shellingham==1.5.4
typer==0.20.0
typing_extensions==4.15.0
